Enable warning 4389 as an error on windows builds.  This will make windows builds more similar to linux/mac, which already treat signed/ unsigned equality comparisons as warnings (and hence errors).
BUG=44471
TEST=none

Review URL: http://codereview.chromium.org/2081007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@48186 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: bd635c097dfc689674d6b21c006850a7b07f23f4

+++ chrome/browser/importer/importer_unittest.cc
@@ -122,7 +122,7 @@ const int kMaxPathSize = 5;
 
 typedef struct {
   const bool in_toolbar;
-  const int path_size;
+  const size_t path_size;
   const wchar_t* path[kMaxPathSize];
   const wchar_t* title;
   const char* url;
@@ -184,7 +184,7 @@ bool FindBookmarkEntry(const ProfileWriter::BookmarkEntry& entry,
         list[i].url == entry.url.spec() &&
         list[i].title == entry.title) {
       bool equal = true;
-      for (int k = 0; k < list[i].path_size; ++k)
+      for (size_t k = 0; k < list[i].path_size; ++k)
         if (list[i].path[k] != entry.path[k]) {
           equal = false;
           break;
